+++ src/App.jsx
@@ -1,14 +1,10 @@
 import './App.css'
-import AdForm from './components/adForm/adForm'
-import UserCard from './components/UserCard'
 import MainPage from './pages/MainPage/MainPage'
 
 function App() {
 
   return (
     <>
-    <AdForm/>
-    <UserCard/>
     </>
   )
 }
