--- src/components/adForm/AdForm.css
@@ -1,92 +1,0 @@
-#form{
-    font-family: system-ui,  sans-serif;
-    width: 30%;
-    margin: auto;
-    box-shadow: rgba(60, 66, 87, 0.12) 0px 7px 14px, rgba(0, 0, 0, 0.12) 0px 3px 6px;
-    padding: 4rem 2.4rem;
-}
-input, select, textarea, #submit {
-    display: block;
-    width: 100%;
-    padding: 6px;
-    box-sizing: border-box;
-    border: unset;
-    border-radius: 4px;
-}
-label{
-    display: block;
-    margin-bottom: 12px;
-}
-input, textarea, select {
-    box-shadow: rgba(60, 66, 87, 0.16) 0px 0px 0px 1px;
-    padding: 8px;
-}
-fieldset{
-    border: none;
-}
-.pad-bottom{
-    padding-bottom: 16px;
-}
-.label{
-    font-size: 16px
-}
-#submit{
-    margin-top: 42px;
-    background-color: #17A043;
-            padding: 8px;
-    color: #fff;
-    transition: box-shadow 1s;
-
-}
-
-#submit:hover{
-    box-shadow: 0px 0px 9px 4px rgba(23,160,67,0.34);}
-textarea {
-    resize: none;
-}
-
-.dropzone {
-    width: 100px;
-    height: 100px;
-    border-radius: 50%;
-    margin-bottom: 16px;
-    cursor: pointer;
-    overflow: hidden;
-    background-size: cover;
-}
-#profile-pic{
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.flex-wrap{
-    display: flex;
-    flex-wrap: wrap;
-}
-
-.category-buttons {
-    width: 100%;
-    display: flex;
-    gap: 10px;
-    flex-wrap: wrap;
-  }
-  
-  .category-button {
-    padding: 10px 20px;
-    background-color: #ddd;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-  }
-  
-  .category-button:hover {
-    background-color: #bbb;
-  }
-  
-  .category-button.selected {
-    background-color: #17A043;
-    color: white;
-    border-color: #17A043;
-  }
